refactor(platform): Migrate store filter to own file
This allows the file to be compiled against all GOOS/GOARCHs.
