+++ machine/platform/register_linux.go
@@ -6,7 +6,6 @@ package platform
 
 import (
 	"context"
-	"fmt"
 	"path/filepath"
 
 	zip "api.zip"
@@ -18,39 +17,6 @@ import (
 	"kraftkit.sh/machine/qemu"
 	"kraftkit.sh/machine/store"
 )
-
-// storePlatformFilter is a mechanism to narrow the results returned from the
-// store which is shared between all platforms.  In this filter, we prefix all
-// requests to the Zip API client with a check for the machine's specification
-// of the platform based on the provided argument.
-func storePlatformFilter(platform Platform) zip.OnBefore {
-	return func(_ context.Context, req zip.ReferenceObject) (any, error) {
-		// If this object is listable, attempt to retrieve from a list from
-		// the store instead.
-		if list, ok := req.(*zip.ObjectList[machinev1alpha1.MachineSpec, machinev1alpha1.MachineStatus]); ok {
-			cached := list.Items
-			list.Items = []zip.Object[machinev1alpha1.MachineSpec, machinev1alpha1.MachineStatus]{}
-
-			for _, machine := range cached {
-				if machine.Spec.Platform != platform.String() {
-					continue
-				}
-
-				list.Items = append(list.Items, machine)
-			}
-			return list, nil
-		}
-
-		// Cast the referenceable object, which we know is a spec-and-status object.
-		obj := req.(*zip.Object[machinev1alpha1.MachineSpec, machinev1alpha1.MachineStatus])
-
-		if obj.Spec.Platform != platform.String() {
-			return nil, fmt.Errorf("machine is not %s instance: %s", platform.String(), obj.Name)
-		}
-
-		return obj, nil
-	}
-}
 
 var qemuV1alpha1Driver = func(ctx context.Context, opts ...any) (machinev1alpha1.MachineService, error) {
 	service, err := qemu.NewMachineV1alpha1Service(ctx, opts...)
